重写用户修改密码逻辑，普通用户修改自己、管理员修改自己、管理员修改普通用户，全部通过

+++ GitCandy.Web/Data/MembershipService.cs
@@ -154,7 +154,7 @@ namespace GitCandy.Data
                     IsAdmin = e.IsAdmin,
                 }).ToArray(),
                 CurrentPage = page,
-                ItemCount = p.TotalCount
+                ItemCount = (Int32)p.TotalCount
             };
         }
 
@@ -369,7 +369,7 @@ namespace GitCandy.Data
                 })
                     .ToArray(),
                 CurrentPage = page,
-                ItemCount = p.TotalCount,
+                ItemCount = (Int32)p.TotalCount,
             };
             return model;
         }
